final version of stationary combustion proxy
